+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 Added
   * CLI command `ocrd-tool`
   * CLI command `workspace find`
+  * CLI command `workspace list-group`
   * CLI command `workspace create`
   * CLI `workspace` flag --no-cache
   * bash library to build wrappers